+++ scripts/ensmallen-release.sh
@@ -23,7 +23,8 @@ if [ "$#" -gt 5 ]; then
 fi
 
 # Make sure that the branch is clean.
-lines=`git diff | wc -l`;
+# Truncate leading whitespaces since wc -l on MacOS adds an extra \t.
+lines=`git diff | wc -l | sed -e 's/^\s*//g'`;
 if [ "$lines" != "0" ]; then
   echo "git diff returned a nonzero result!";
   echo "";
@@ -131,6 +132,15 @@ git add CONTRIBUTING.md;
 git add HISTORY.md;
 git commit -m "Update and release version $MAJOR.$MINOR.$PATCH.";
 
+changelog_str=`cat HISTORY.md |\
+    awk '/^### /{f=0} /^### ensmallen '"$MAJOR"'.'"$MINOR"'.'"$PATCH"': "'"$version_name"'"/{f=1} f{print}' |\
+    grep -v '^#' |\
+    tr '\n' '!' |\
+    sed -e 's/!  [ ]*/ /g' |\
+    tr '!' '\n'`;
+echo "Changelog string:"
+echo "$changelog_str"
+
 # Add one more commit to create the new HISTORY block.
 echo "### ensmallen ?.??.?: \"???\"" > HISTORY.md.new;
 echo "###### ????-??-??" >> HISTORY.md.new;
@@ -148,10 +158,13 @@ git push --set-upstream $github_user release-$MAJOR.$MINOR.$PATCH;
 hub pull-request \
     -b mlpack:master \
     -h $github_user:release-$MAJOR.$MINOR.$PATCH \
-    -m "Release version $MAJOR.$MINOR.$PATCH" \
+    -m "Release version $MAJOR.$MINOR.$PATCH: \"$version_name\"" \
     -m "This automatically-generated pull request adds the commits necessary to make the $MAJOR.$MINOR.$PATCH release." \
     -m "Once the PR is merged, mlpack-bot will tag the release as HEAD~1 (so that it doesn't include the new HISTORY block) and publish it." \
     -m "Or, well, hopefully that will happen someday." \
+    -m "When you merge this PR, be sure to merge it using a *rebase*." \
+    -m "### Changelog" \
+    -m "$changelog_str" \
     -l "t: release"
 
 echo "";